+++ packages/component/src/providers/ReducedMotion/useShouldReduceMotion.ts
@@ -1,0 +1,5 @@
+import useContext from './private/useContext';
+
+export default function useShouldReduceMotion(): ReturnType<typeof useContext>['shouldReduceMotionState'] {
+  return useContext().shouldReduceMotionState;
+}